refactor(gui-agent): consolidate architecture and remove RefactoredOperator
- Merge RefactoredOperator functionality into BrowserOperator base class
- Simplify inheritance hierarchy, and clean up type definitions.
- BREAKING CHANGE: RefactoredOperator removed, API simplified, BrowserOperator refactored
