+++ multimodal/gui-agent/operator-browser/package.json
@@ -37,8 +37,6 @@
     "@agent-infra/browser": "0.1.1",
     "@agent-infra/logger": "0.0.2-beta.2",
     "@agent-infra/puppeteer-enhance": "0.1.6",
-    "@ui-tars/sdk": "1.2.3",
-    "@ui-tars/shared": "1.2.3",
     "@gui-agent/shared": "workspace:*"
   },
   "devDependencies": {
