+++ multimodal/gui-agent/operator-browser/src/index.ts
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 Bytedance, Inc. and its affiliates.
  * SPDX-License-Identifier: Apache-2.0
  */
-export { BrowserOperator, DefaultBrowserOperator } from './browser-operator';
+export { BrowserOperator } from './browser-operator';
 export { LocalBrowserOperator } from './LocalBrowserOperator';
 export { RemoteBrowserOperator } from './RemoteBrowserOperator';
 export * from './types';
