+++ multimodal/gui-agent/operator-browser/src/types.ts
@@ -3,12 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { Logger } from '@agent-infra/logger';
-import type { Page, BrowserType, BaseBrowser } from '@agent-infra/browser';
-import type { ScreenshotOutput, ExecuteParams } from '@ui-tars/sdk/core';
-
-export { StatusEnum } from '@ui-tars/sdk';
-export type { Page, ScreenshotOutput, ExecuteParams };
-export type ParsedPrediction = ExecuteParams['parsedPrediction'];
+import type { BrowserType, BaseBrowser } from '@agent-infra/browser';
 
 /**
  * Search engine options
@@ -62,22 +57,6 @@ export interface BrowserOperatorOptions {
    * @default true
    */
   showWaterFlow?: boolean;
-
-  /**
-   * Callback triggered when an operator action is performed
-   * @deprecated Will be removed when `@ui-tars/sdk` supports hooks natively
-   */
-  onOperatorAction?: (prediction: ParsedPrediction) => Promise<void>;
-
-  /**
-   * Callback triggered when a screenshot is taken
-   */
-  onScreenshot?: (screenshot: ScreenshotOutput, page: Page) => Promise<void>;
-
-  /**
-   * Callback triggered when a final answer is received
-   */
-  onFinalAnswer?: (finalAnswer: string) => Promise<void>;
 }
 
 export interface LocalBrowserOperatorOptions {
